move to trace: autopatch,call_context
