+++ weave/conftest.py
@@ -26,10 +26,11 @@ from weave.trace_server import (
     trace_server_interface as tsi,
 )
 
-from . import autopatch, environment, logs
+from . import environment, logs
 from .tests import fixture_fakewandb
 from .tests.trace.trace_server_clickhouse_conftest import *
 from .tests.wandb_system_tests_conftest import *
+from .trace import autopatch
 
 logs.configure_logger()
